decompose selector
