+++ package.json
@@ -26,22 +26,24 @@
         "src/index.js"
     ],
     "dependencies": {
-        "lodash.isequal": "^4.5.0"
+        "lodash.isequal": "^4.5.0",
+        "lodash.isfunction": "^3.0.9",
+        "memoize-one": "^5.1.1"
     },
     "devDependencies": {
         "babel-jest": "^23.6.0",
         "babel-preset-react-native": "^4.0.1",
         "enzyme": "^3.7.0",
         "enzyme-adapter-react-16": "^1.7.0",
-        "enzyme-to-json": "^3.3.5",
+        "enzyme-to-json": "^3.4.0",
         "eslint-config-ls-react": "https://github.com/lawnstarter/eslint-config-ls-react#3.0.0",
         "husky": "^2.4.0",
         "jest": "^23.6.0",
         "prettier": "^1.15.3",
         "pretty-quick": "^1.8.0",
         "prop-types": "^15.6.2",
-        "react": "16.6.1",
-        "react-dom": "^16.6.1",
+        "react": "^16.9.0",
+        "react-dom": "^16.9.0",
         "react-native": "0.57.7",
         "react-test-renderer": "^16.6.1"
     },
