+++ src/index.js
@@ -1,127 +1,32 @@
+// eslint-disable-next-line no-unused-vars
 import React, { PureComponent } from 'react';
-import {
-    ColorPropType,
-    Keyboard,
-    Modal,
-    Picker,
-    Platform,
-    StyleSheet,
-    Text,
-    TextInput,
-    TouchableOpacity,
-    TouchableWithoutFeedback,
-    View,
-} from 'react-native';
+import { ColorPropType, Keyboard, Platform } from 'react-native';
+
 import PropTypes from 'prop-types';
 import isEqual from 'lodash.isequal';
 
+import styles from './styles';
+
+import {
+    handlePlaceholder,
+    renderIOS,
+    getSelectedItem,
+    renderAndroidHeadless,
+    renderAndroidNativePickerStyle,
+} from './components';
+
+export const defaultStyles = styles;
+
 export default class RNPickerSelect extends PureComponent {
-    static propTypes = {
-        onValueChange: PropTypes.func.isRequired,
-        items: PropTypes.arrayOf(
-            PropTypes.shape({
-                label: PropTypes.string.isRequired,
-                value: PropTypes.any.isRequired,
-                key: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-                color: ColorPropType,
-            })
-        ).isRequired,
-        value: PropTypes.any, // eslint-disable-line react/forbid-prop-types
-        placeholder: PropTypes.shape({
-            label: PropTypes.string,
-            value: PropTypes.any,
-            key: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-            color: ColorPropType,
-        }),
-        disabled: PropTypes.bool,
-        itemKey: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-        style: PropTypes.shape({}),
-        children: PropTypes.any, // eslint-disable-line react/forbid-prop-types
-        placeholderTextColor: ColorPropType, // deprecated
-        useNativeAndroidPickerStyle: PropTypes.bool,
-
-        // Custom Modal props (iOS only)
-        hideDoneBar: PropTypes.bool, // deprecated
-        doneText: PropTypes.string,
-        onDonePress: PropTypes.func,
-        onUpArrow: PropTypes.func,
-        onDownArrow: PropTypes.func,
-        onOpen: PropTypes.func,
-        onClose: PropTypes.func,
-
-        // Modal props (iOS only)
-        modalProps: PropTypes.shape({}),
-
-        // TextInput props (iOS only)
-        textInputProps: PropTypes.shape({}),
-
-        // Picker props
-        pickerProps: PropTypes.shape({}),
-
-        // Custom Icon
-        Icon: PropTypes.func,
-        InputAccessoryView: PropTypes.func,
-    };
-
-    static defaultProps = {
-        value: undefined,
-        placeholder: {
-            label: 'Select an item...',
-            value: null,
-            color: '#9EA0A4',
-        },
-        disabled: false,
-        itemKey: null,
-        style: {},
-        children: null,
-        placeholderTextColor: '#C7C7CD', // deprecated
-        useNativeAndroidPickerStyle: true,
-        hideDoneBar: false, // deprecated
-        doneText: 'Done',
-        onDonePress: null,
-        onUpArrow: null,
-        onDownArrow: null,
-        onOpen: null,
-        onClose: null,
-        modalProps: {},
-        textInputProps: {},
-        pickerProps: {},
-        Icon: null,
-        InputAccessoryView: null,
-    };
-
-    static handlePlaceholder({ placeholder }) {
-        if (isEqual(placeholder, {})) {
-            return [];
-        }
-        return [placeholder];
-    }
-
-    static getSelectedItem({ items, key, value }) {
-        let idx = items.findIndex((item) => {
-            if (item.key && key) {
-                return isEqual(item.key, key);
-            }
-            return isEqual(item.value, value);
-        });
-        if (idx === -1) {
-            idx = 0;
-        }
-        return {
-            selectedItem: items[idx] || {},
-            idx,
-        };
-    }
-
     static getDerivedStateFromProps(nextProps, prevState) {
         // update items if items or placeholder prop changes
-        const items = RNPickerSelect.handlePlaceholder({
+        const items = handlePlaceholder({
             placeholder: nextProps.placeholder,
         }).concat(nextProps.items);
         const itemsChanged = !isEqual(prevState.items, items);
 
         // update selectedItem if value prop is defined and differs from currently selected item
-        const { selectedItem, idx } = RNPickerSelect.getSelectedItem({
+        const { selectedItem, idx } = getSelectedItem({
             items,
             key: nextProps.itemKey,
             value: nextProps.value,
@@ -146,11 +51,11 @@ export default class RNPickerSelect extends PureComponent {
     constructor(props) {
         super(props);
 
-        const items = RNPickerSelect.handlePlaceholder({
+        const items = handlePlaceholder({
             placeholder: this.props.placeholder,
         }).concat(this.props.items);
 
-        const { selectedItem } = RNPickerSelect.getSelectedItem({
+        const { selectedItem } = getSelectedItem({
             items,
             key: this.props.itemKey,
             value: this.props.value,
@@ -163,30 +68,21 @@ export default class RNPickerSelect extends PureComponent {
             animationType: undefined,
             orientation: 'portrait',
         };
-
-        this.onUpArrow = this.onUpArrow.bind(this);
-        this.onDownArrow = this.onDownArrow.bind(this);
-        this.onValueChange = this.onValueChange.bind(this);
-        this.onOrientationChange = this.onOrientationChange.bind(this);
-        this.setInputRef = this.setInputRef.bind(this);
-        this.togglePicker = this.togglePicker.bind(this);
-        this.triggerDoneCallback = this.triggerDoneCallback.bind(this);
-        this.renderInputAccessoryView = this.renderInputAccessoryView.bind(this);
     }
 
-    onUpArrow() {
+    onUpArrow = () => {
         const { onUpArrow } = this.props;
 
         this.togglePicker(false, onUpArrow);
-    }
+    };
 
-    onDownArrow() {
+    onDownArrow = () => {
         const { onDownArrow } = this.props;
 
         this.togglePicker(false, onDownArrow);
-    }
+    };
 
-    onValueChange(value, index) {
+    onValueChange = (value, index) => {
         const { onValueChange } = this.props;
 
         onValueChange(value, index);
@@ -196,32 +92,15 @@ export default class RNPickerSelect extends PureComponent {
                 selectedItem: prevState.items[index],
             };
         });
-    }
+    };
 
-    onOrientationChange({ nativeEvent }) {
+    onOrientationChange = ({ nativeEvent }) => {
         this.setState({
             orientation: nativeEvent.orientation,
         });
-    }
+    };
 
-    setInputRef(ref) {
-        this.inputRef = ref;
-    }
-
-    getPlaceholderStyle() {
-        const { placeholder, placeholderTextColor, style } = this.props;
-
-        if (!isEqual(placeholder, {}) && this.state.selectedItem.label === placeholder.label) {
-            return {
-                ...defaultStyles.placeholder,
-                color: placeholderTextColor, // deprecated
-                ...style.placeholder,
-            };
-        }
-        return {};
-    }
-
-    triggerOpenCloseCallbacks() {
+    triggerOpenCloseCallbacks = () => {
         const { onOpen, onClose } = this.props;
 
         if (!this.state.showPicker && onOpen) {
@@ -231,16 +110,16 @@ export default class RNPickerSelect extends PureComponent {
         if (this.state.showPicker && onClose) {
             onClose();
         }
-    }
+    };
 
-    triggerDoneCallback() {
+    triggerDoneCallback = () => {
         const { hideDoneBar, onDonePress } = this.props;
         if (!hideDoneBar && onDonePress) {
             onDonePress();
         }
-    }
+    };
 
-    togglePicker(animate = false, postToggleCallback) {
+    togglePicker = (animate = false, postToggleCallback) => {
         const { modalProps, disabled } = this.props;
 
         if (disabled) {
@@ -252,6 +131,7 @@ export default class RNPickerSelect extends PureComponent {
         }
 
         const animationType =
+            // eslint-disable-next-line react/prop-types
             modalProps && modalProps.animationType ? modalProps.animationType : 'slide';
 
         this.triggerOpenCloseCallbacks();
@@ -269,317 +149,115 @@ export default class RNPickerSelect extends PureComponent {
                 }
             }
         );
-    }
-
-    renderPickerItems() {
-        return this.state.items.map((item) => {
-            return (
-                <Picker.Item
-                    label={item.label}
-                    value={item.value}
-                    key={item.key || item.label}
-                    color={item.color}
-                />
-            );
-        });
-    }
-
-    renderInputAccessoryView() {
-        const {
-            InputAccessoryView,
-            doneText,
-            hideDoneBar,
-            onUpArrow,
-            onDownArrow,
-            style,
-        } = this.props;
-
-        // deprecated
-        if (hideDoneBar) {
-            return null;
-        }
-
-        if (InputAccessoryView) {
-            return <InputAccessoryView testID="custom_input_accessory_view" />;
-        }
-
-        return (
-            <View
-                style={[defaultStyles.modalViewMiddle, style.modalViewMiddle]}
-                testID="input_accessory_view"
-            >
-                <View style={[defaultStyles.chevronContainer, style.chevronContainer]}>
-                    <TouchableOpacity
-                        activeOpacity={onUpArrow ? 0.5 : 1}
-                        onPress={onUpArrow ? this.onUpArrow : null}
-                    >
-                        <View
-                            style={[
-                                defaultStyles.chevron,
-                                style.chevron,
-                                defaultStyles.chevronUp,
-                                style.chevronUp,
-                                onUpArrow ? [defaultStyles.chevronActive, style.chevronActive] : {},
-                            ]}
-                        />
-                    </TouchableOpacity>
-                    <TouchableOpacity
-                        activeOpacity={onDownArrow ? 0.5 : 1}
-                        onPress={onDownArrow ? this.onDownArrow : null}
-                    >
-                        <View
-                            style={[
-                                defaultStyles.chevron,
-                                style.chevron,
-                                defaultStyles.chevronDown,
-                                style.chevronDown,
-                                onDownArrow
-                                    ? [defaultStyles.chevronActive, style.chevronActive]
-                                    : {},
-                            ]}
-                        />
-                    </TouchableOpacity>
-                </View>
-                <TouchableWithoutFeedback
-                    onPress={() => {
-                        this.togglePicker(true);
-                    }}
-                    hitSlop={{ top: 4, right: 4, bottom: 4, left: 4 }}
-                    testID="done_button"
-                >
-                    <View testID="needed_for_touchable">
-                        <Text style={[defaultStyles.done, style.done]}>{doneText}</Text>
-                    </View>
-                </TouchableWithoutFeedback>
-            </View>
-        );
-    }
-
-    renderIcon() {
-        const { style, Icon } = this.props;
-
-        if (!Icon) {
-            return null;
-        }
-
-        return (
-            <View
-                testID="icon_container"
-                style={[defaultStyles.iconContainer, style.iconContainer]}
-            >
-                <Icon testID="icon" />
-            </View>
-        );
-    }
-
-    renderTextInputOrChildren() {
-        const { children, style, textInputProps } = this.props;
-        const containerStyle =
-            Platform.OS === 'ios' ? style.inputIOSContainer : style.inputAndroidContainer;
-
-        if (children) {
-            return (
-                <View pointerEvents="box-only" style={containerStyle}>
-                    {children}
-                </View>
-            );
-        }
-
-        return (
-            <View pointerEvents="box-only" style={containerStyle}>
-                <TextInput
-                    style={[
-                        Platform.OS === 'ios' ? style.inputIOS : style.inputAndroid,
-                        this.getPlaceholderStyle(),
-                    ]}
-                    value={this.state.selectedItem.label}
-                    ref={this.setInputRef}
-                    editable={false}
-                    {...textInputProps}
-                />
-                {this.renderIcon()}
-            </View>
-        );
-    }
-
-    renderIOS() {
-        const { style, modalProps, pickerProps } = this.props;
-
-        return (
-            <View style={[defaultStyles.viewContainer, style.viewContainer]}>
-                <TouchableWithoutFeedback
-                    onPress={() => {
-                        this.togglePicker(true);
-                    }}
-                    testID="ios_touchable_wrapper"
-                >
-                    {this.renderTextInputOrChildren()}
-                </TouchableWithoutFeedback>
-                <Modal
-                    testID="ios_modal"
-                    visible={this.state.showPicker}
-                    transparent
-                    animationType={this.state.animationType}
-                    supportedOrientations={['portrait', 'landscape']}
-                    onDismiss={this.triggerDoneCallback}
-                    onOrientationChange={this.onOrientationChange}
-                    {...modalProps}
-                >
-                    <TouchableOpacity
-                        style={[defaultStyles.modalViewTop, style.modalViewTop]}
-                        testID="ios_modal_top"
-                        onPress={() => {
-                            this.togglePicker(true);
-                        }}
-                    />
-                    {this.renderInputAccessoryView()}
-                    <View
-                        style={[
-                            defaultStyles.modalViewBottom,
-                            { height: this.state.orientation === 'portrait' ? 215 : 162 },
-                            style.modalViewBottom,
-                        ]}
-                    >
-                        <Picker
-                            testID="ios_picker"
-                            onValueChange={this.onValueChange}
-                            selectedValue={this.state.selectedItem.value}
-                            {...pickerProps}
-                        >
-                            {this.renderPickerItems()}
-                        </Picker>
-                    </View>
-                </Modal>
-            </View>
-        );
-    }
-
-    renderAndroidHeadless() {
-        const { disabled, Icon, style, pickerProps } = this.props;
-
-        return (
-            <View style={style.headlessAndroidContainer}>
-                {this.renderTextInputOrChildren()}
-                <Picker
-                    style={[
-                        Icon ? { backgroundColor: 'transparent' } : {}, // to hide native icon
-                        defaultStyles.headlessAndroidPicker,
-                        style.headlessAndroidPicker,
-                    ]}
-                    testID="android_picker_headless"
-                    enabled={!disabled}
-                    onValueChange={this.onValueChange}
-                    selectedValue={this.state.selectedItem.value}
-                    {...pickerProps}
-                >
-                    {this.renderPickerItems()}
-                </Picker>
-            </View>
-        );
-    }
-
-    renderAndroidNativePickerStyle() {
-        const { disabled, Icon, style, pickerProps } = this.props;
-
-        return (
-            <View style={[defaultStyles.viewContainer, style.viewContainer]}>
-                <Picker
-                    style={[
-                        Icon ? { backgroundColor: 'transparent' } : {}, // to hide native icon
-                        style.inputAndroid,
-                        this.getPlaceholderStyle(),
-                    ]}
-                    testID="android_picker"
-                    enabled={!disabled}
-                    onValueChange={this.onValueChange}
-                    selectedValue={this.state.selectedItem.value}
-                    {...pickerProps}
-                >
-                    {this.renderPickerItems()}
-                </Picker>
-                {this.renderIcon()}
-            </View>
-        );
-    }
+    };
 
     render() {
         const { children, useNativeAndroidPickerStyle } = this.props;
+        const { items, showPicker, animationType, selectedItem, orientation } = this.state;
+
+        const props = {
+            ...this.props,
+            orientation,
+            selectedItem,
+            animationType,
+            showPicker,
+            items,
+            onValueChange: this.onValueChange,
+            togglePicker: (value) => {
+                this.togglePicker(value);
+            },
+            onOrientationChange: this.onOrientationChange,
+            triggerDoneCallback: this.triggerDoneCallback,
+            onUpArrow: this.onUpArrow,
+            onDownArrow: this.onDownArrow,
+        };
 
         if (Platform.OS === 'ios') {
-            return this.renderIOS();
+            return renderIOS(props);
         }
 
         if (children || !useNativeAndroidPickerStyle) {
-            return this.renderAndroidHeadless();
+            return renderAndroidHeadless(props);
         }
 
-        return this.renderAndroidNativePickerStyle();
+        return renderAndroidNativePickerStyle(props);
     }
 }
 
-export const defaultStyles = StyleSheet.create({
-    viewContainer: {
-        alignSelf: 'stretch',
-    },
-    iconContainer: {
-        position: 'absolute',
-        right: 0,
-    },
-    modalViewTop: {
-        flex: 1,
-    },
-    modalViewMiddle: {
-        height: 44,
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        paddingHorizontal: 10,
-        backgroundColor: '#EFF1F2',
-        borderTopWidth: 0.5,
-        borderTopColor: '#919498',
-        zIndex: 2,
-    },
-    chevronContainer: {
-        flexDirection: 'row',
-    },
-    chevron: {
-        width: 15,
-        height: 15,
-        backgroundColor: 'transparent',
-        borderColor: '#D0D4DB',
-        borderTopWidth: 1.5,
-        borderRightWidth: 1.5,
-    },
-    chevronUp: {
-        marginLeft: 11,
-        transform: [{ translateY: 4 }, { rotate: '-45deg' }],
-    },
-    chevronDown: {
-        marginLeft: 22,
-        transform: [{ translateY: -5 }, { rotate: '135deg' }],
-    },
-    chevronActive: {
-        borderColor: '#007AFE',
-    },
-    done: {
-        color: '#007AFE',
-        fontWeight: 'bold',
-        fontSize: 15,
-        paddingTop: 1,
-        paddingRight: 2,
-    },
-    modalViewBottom: {
-        justifyContent: 'center',
-        backgroundColor: '#D0D4DB',
-    },
+RNPickerSelect.propTypes = {
+    onValueChange: PropTypes.func.isRequired,
+    items: PropTypes.arrayOf(
+        PropTypes.shape({
+            label: PropTypes.string.isRequired,
+            value: PropTypes.any.isRequired,
+            key: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+            color: ColorPropType,
+        })
+    ).isRequired,
+    value: PropTypes.any, // eslint-disable-line react/forbid-prop-types
+    placeholder: PropTypes.shape({
+        label: PropTypes.string,
+        value: PropTypes.any,
+        key: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+        color: ColorPropType,
+    }),
+    disabled: PropTypes.bool,
+    itemKey: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+    style: PropTypes.shape({}),
+    children: PropTypes.any, // eslint-disable-line react/forbid-prop-types
+    // eslint-disable-next-line react/no-unused-prop-types
+    placeholderTextColor: ColorPropType, // deprecated
+    useNativeAndroidPickerStyle: PropTypes.bool,
+
+    // Custom Modal props (iOS only)
+    hideDoneBar: PropTypes.bool, // deprecated
+    // eslint-disable-next-line react/no-unused-prop-types
+    doneText: PropTypes.string,
+    onDonePress: PropTypes.func,
+    onUpArrow: PropTypes.func,
+    onDownArrow: PropTypes.func,
+    onOpen: PropTypes.func,
+    onClose: PropTypes.func,
+
+    // Modal props (iOS only)
+    modalProps: PropTypes.shape({}),
+
+    // TextInput props (iOS only)
+    textInputProps: PropTypes.shape({}),
+
+    // Picker props
+    pickerProps: PropTypes.shape({}),
+
+    // Custom Icon
+    // eslint-disable-next-line react/no-unused-prop-types
+    Icon: PropTypes.func,
+    // eslint-disable-next-line react/no-unused-prop-types
+    InputAccessoryView: PropTypes.func,
+};
+
+RNPickerSelect.defaultProps = {
+    value: undefined,
     placeholder: {
-        color: '#C7C7CD',
+        label: 'Select an item...',
+        value: null,
+        color: '#9EA0A4',
     },
-    headlessAndroidPicker: {
-        position: 'absolute',
-        width: '100%',
-        height: '100%',
-        color: 'transparent',
-        opacity: 0,
-    },
-});
+    disabled: false,
+    itemKey: null,
+    style: {},
+    children: null,
+    placeholderTextColor: '#C7C7CD', // deprecated
+    useNativeAndroidPickerStyle: true,
+    hideDoneBar: false, // deprecated
+    doneText: 'Done',
+    onDonePress: null,
+    onUpArrow: null,
+    onDownArrow: null,
+    onOpen: null,
+    onClose: null,
+    modalProps: {},
+    textInputProps: {},
+    pickerProps: {},
+    Icon: null,
+    InputAccessoryView: null,
+};
