+++ src/styles.js
@@ -1,0 +1,70 @@
+import { StyleSheet } from 'react-native';
+
+const defaultStyles = StyleSheet.create({
+    viewContainer: {
+        alignSelf: 'stretch',
+    },
+    iconContainer: {
+        position: 'absolute',
+        right: 0,
+    },
+    modalViewTop: {
+        flex: 1,
+    },
+    modalViewMiddle: {
+        height: 44,
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        paddingHorizontal: 10,
+        backgroundColor: '#EFF1F2',
+        borderTopWidth: 0.5,
+        borderTopColor: '#919498',
+        zIndex: 2,
+    },
+    chevronContainer: {
+        flexDirection: 'row',
+    },
+    chevron: {
+        width: 15,
+        height: 15,
+        backgroundColor: 'transparent',
+        borderColor: '#D0D4DB',
+        borderTopWidth: 1.5,
+        borderRightWidth: 1.5,
+    },
+    chevronUp: {
+        marginLeft: 11,
+        transform: [{ translateY: 4 }, { rotate: '-45deg' }],
+    },
+    chevronDown: {
+        marginLeft: 22,
+        transform: [{ translateY: -5 }, { rotate: '135deg' }],
+    },
+    chevronActive: {
+        borderColor: '#007AFE',
+    },
+    done: {
+        color: '#007AFE',
+        fontWeight: 'bold',
+        fontSize: 15,
+        paddingTop: 1,
+        paddingRight: 2,
+    },
+    modalViewBottom: {
+        justifyContent: 'center',
+        backgroundColor: '#D0D4DB',
+    },
+    placeholder: {
+        color: '#C7C7CD',
+    },
+    headlessAndroidPicker: {
+        position: 'absolute',
+        width: '100%',
+        height: '100%',
+        color: 'transparent',
+        opacity: 0,
+    },
+});
+
+export default defaultStyles;
